requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-environ==0.9.0
Pillow==9.4.0
psycopg2==2.9.5
Pyrogram==2.0.100
pyromod==1.5
TgCrypto==1.2.5

requests==2.28.1